requirements.txt
arabic-reshaper==3.0.0
asgiref==3.8.1
asn1crypto==1.5.1
Brotli==1.1.0
certifi==2024.7.4
cffi==1.17.0
chardet==5.2.0
charset-normalizer==3.3.2
class-registry==2.1.2
click==8.1.7
colorama==0.4.6
cryptography==43.0.0
cssselect2==0.7.0
diff-match-patch==20230430
distlib==0.3.8
dj-database-url==2.2.0
Django==5.0.8
django-environ==0.11.2
django-import-export==4.1.1
django-jazzmin==3.0.0
django-pandas==0.6.7
django-pdfkit==0.3.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
et-xmlfile==1.1.0
filelock==3.15.4
filters==1.3.2
gunicorn==23.0.0
h11==0.14.0
html5lib==1.1
idna==3.7
lxml==5.3.0
Markdown==3.6
mysqlclient==2.2.4
ndg-httpsclient==0.5.1
numpy==2.0.1
openpyxl==3.1.5
oscrypto==1.3.0
packaging==24.1
pandas==2.2.2
pdfkit==0.6.0
pillow==10.4.0
platformdirs==4.2.2
psycopg2==2.9.9
psycopg2-binary==2.9.9
pyasn1==0.6.0
pycparser==2.22
pyHanko==0.25.1
pyhanko-certvalidator==0.26.3
PyJWT==2.9.0
pyOpenSSL==24.2.1
pypdf==4.3.1
pypng==0.20220715.0
python-bidi==0.6.0
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.2
qrcode==7.4.2
regex==2024.7.24
reportlab==4.2.2
requests==2.32.3
rest-framework-simplejwt==0.0.2
simplejwt==2.0.1
six==1.16.0
sqlparse==0.5.1
svglib==1.5.1
tablib==3.5.0
tinycss2==1.3.0
typing==3.7.4.3
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
uritools==4.0.3
urllib3==2.2.2
uvicorn==0.30.6
virtualenv==20.26.3
webencodings==0.5.1
whitenoise==6.7.0
xhtml2pdf==0.2.16
xlrd==2.0.1


